+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Updated to newest pytorch lightning;
 - Semantic Segmentation model updated for multi class;
 - New image callback for multi class semantic segmentation;
+- Added improvements for OneCycleLR scheduler (automatic calculation of steps_per_epoch);
 
 # Version 0.17.0